[feat]: finish the basic info_proc

+++ Draft/Inc/bsp/bsp_spi.h
@@ -1,8 +1,9 @@
 #ifndef BSP_SPI_H
 #define BSP_SPI_H
 #include "struct_typedef.h"
+#include "main.h"
 
-extern void SPI1_DMA_init(uint64_t tx_buf, uint64_t rx_buf, uint16_t num);
-extern void SPI1_DMA_enable(uint64_t tx_buf, uint64_t rx_buf, uint16_t ndtr);
+extern void SPI1_DMA_init(uint32_t tx_buf, uint32_t rx_buf, uint16_t num);
+extern void SPI1_DMA_enable(uint32_t tx_buf, uint32_t rx_buf, uint16_t ndtr);
 
 #endif
